Projects/websites_with_flask/92(image_to_colour_palette_generator)/requirements.txt
Flask==2.3.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
Jinja2==3.1.2
werkzeug==2.3.0
pillow==10.0.0
colorthief==0.2.1
webcolors==1.13